requirements.txt
pywinauto==0.6.8
pywin32==306
mouse==0.7.1
keyboard==0.13.5
pyautogui==0.9.53
mss==7.0.1
fuzzywuzzy==0.18.0
nltk==3.8.1
python-redmine==2.4.0
selenium==4.8.0
Pillow==9.4.0
python-dateutil==2.8.2
openpyxl==3.1.2
pandas==2.0.0
docxtpl==0.16.6
python-redmine==2.4.0
xlrd==2.0.1
py7zr==0.20.5